added byprice search, need to add screen path

+++ src/main/java/com/revature/ecommerce_cli/DAO/ProductDAO.java
@@ -178,8 +178,8 @@ public class ProductDAO implements CrudDAO<Product> {
             String sql = "select * from products where price between ? and ?";
 
             try (PreparedStatement ps = conn.prepareStatement(sql)) {
-                ps.setInt(1, first);
-                ps.setInt(2, second);
+                ps.setDouble(1, first);
+                ps.setDouble(2, second);
                 try(ResultSet rs = ps.executeQuery()) {
                     while(rs.next()) {
                         retArray.add(getFromResultSet(rs));
